requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
supabase>=2.3.0
openai>=1.12.0
httpx>=0.26.0
